fix: resolve compiler warnings and improve pre-commit hooks
What I've done:
- Fixed signed/unsigned comparison warning in Lexer.h.
- Suppressed third-party library warnings in CMakeLists.txt using _SILENCE_NONFLOATING_COMPLEX_DEPRECATION_WARNING and _CRT_SECURE_NO_WARNINGS.
- Added -Wno-unused-parameter and /wd4100 to suppress unavoidable warnings from external headers.
- Fixed 'unused-but-set-variable' warning in MLIRGen.cpp.
- Improved pre-commit hook to catch clang-tidy failures more reliably using a loop.
- Fixed markdownlint issues in docs/compilation.md (applied via pre-commit hook).

Verification:
- Project builds cleanly with -Wall -Wextra -Werror.
- Pre-commit hooks passed successfully.

+++ include/toy/Lexer.h
@@ -66,7 +66,7 @@ class Lexer {
   int curCol = 0;
   Location lastLoc = {1, 0};
   const std::string source;
-  int curPos = 0;
+  size_t curPos = 0;
 
   char nextChar() {
     if (curPos >= source.size()) return EOF;
